Create a dedicated `SystemActor` type
